Add files via upload
This code will generate the pie chart of all nucleotide bases present in the input DNA sequence. It will also print the length of the sequence,
GC content and our input DNA. This code can take multiple inputs from users and we want to stop the code at some point then just we need to write a stop that will terminate the process.